xc2par: Refactor NetlistGraph* to IntermediateGraph*
The existing NetlistGraph is not very convenient because it essentially
erases information about which specific node variant is present in a
given slot. This makes manipulating the NetlistGraph essentially
"dynamically typed" and needs checking everywhere. The existing
NetlistMacrocell is going to be modified into a "statically typed" graph
data structure, so rename the existing NetlistGraph to IntermediateGraph
to make it more clear what it is doing.